LinGui: add video quality slider granularity preference. I didn't have an easy way to stash doubles in a combo box, so this was more invasive that planned. But now I have some additional flexibitlity.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@2154 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ gtk/src/x264handler.c
@@ -254,7 +254,7 @@ x264_update_combo(signal_user_data_t *ud, const gchar *name, const gchar *val)
 	GtkTreeModel *store;
 	GtkTreeIter iter;
 	gchar *shortOpt;
-	gint ivalue;
+	gdouble ivalue;
 	gboolean foundit = FALSE;
 	GtkWidget *widget;
